# perf(client): optimize CheckCometError with structured error matching